// pslpwcl/src/app/modules/pslpapi/model/incontriAppuntamentiRequest.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CallInfo } from './callInfo';
import { DettaglioIncontro } from './dettaglioIncontro';

export interface IncontriAppuntamentiRequest { 
    callInfo?: CallInfo;
    dataIncontro?: Date;
    dettaglioIncontroPrecedente?: DettaglioIncontro;
    durataEffettiva?: number;
    flgOperatore?: string;
    idModalitaErogazione?: number;
    idMotivazione?: number;
    idOperatore?: string;
    idSilLavAnagrafica?: number;
    idSilTCpi?: number;
    idSilwebTIncServizi?: number;
    idSilwebTPianifLabor?: number;
    idSilwebTSedeEnte?: number;
    idStatoIncon?: string;
    note?: string;
    operazione?: string;
    oraIncontro?: string;
}